simplify some declarations
There were multiple declarations in pattern of `integer(1)` which
better are declared as `integer` only.  An accidental conversion
from INTEGER(4) to INTEGER(1), as the compilation with gfortran's
option -Wall describes as possible, then is less likely.  This
equally is a point suggested earlier by Vincent.

File I/O, validity test, provision of a solution for a grid read
from a file (check with diff -s, test_in_01.txt, test_out_02.txt)
do not reveal a change in the functionality.

+++ app/main.f90
@@ -25,10 +25,10 @@ program main
 
     implicit none
     ! Variables locales :
-    integer(1), dimension(1:9, 1:9) :: grille    ! (ligne,colonne)
+    integer, dimension(1:9, 1:9) :: grille    ! (ligne,colonne)
     real(kind=dp)    :: Debut,Fin    ! Pour mesurer la durée de calcul
-    integer(1)    :: choix
-    integer(1)    :: nvides    ! Nombre de cases à vider
+    integer    :: choix
+    integer    :: nvides    ! Nombre de cases à vider
     character(50) :: fichier    ! Nom du fichier .txt
 
     ! Initialisation du générateur de nombres pseudo-aléatoires :
